+++ src/containers/BooksForm.js
@@ -1,0 +1,41 @@
+import React from 'react';
+
+const categories = [
+  'Action',
+  'Biography',
+  'History',
+  'Horror',
+  'Kids',
+  'Learning',
+  'Sci-Fi',
+];
+
+class BooksForm extends React.Component {
+  constructor(props) {
+    super(props);
+    this.state = {
+      title: '',
+      category: '',
+    };
+  }
+
+  render() {
+    const { title, category } = this.state;
+    return (
+      <form>
+        <input id="title" value={title} />
+        <select value={category}>
+          {
+            categories.map(cat => (
+              <option key={cat}>
+                {cat}
+              </option>
+            ))
+          }
+        </select>
+      </form>
+    );
+  }
+}
+
+export default BooksForm;
