+++ src/index.js
@@ -2,10 +2,9 @@ import React from 'react';
 import { Provider } from 'react-redux';
 import ReactDOM from 'react-dom';
 import { createStore } from 'redux';
-import bookReducer from './reducers/books';
+import rootReducer from './reducers/index';
 import './index.css';
 import App from './components/App';
-import * as serviceWorker from './serviceWorker';
 
 const initialState = {
   books: [
@@ -22,7 +21,7 @@ const initialState = {
   ],
 };
 
-const store = createStore(bookReducer, initialState);
+const store = createStore(rootReducer, initialState);
 
 ReactDOM.render(
   <React.StrictMode>
@@ -32,8 +31,3 @@ ReactDOM.render(
   </React.StrictMode>,
   document.getElementById('root'),
 );
-
-// If you want your app to work offline and load faster, you can change
-// unregister() to register() below. Note this comes with some pitfalls.
-// Learn more about service workers: https://bit.ly/CRA-PWA
-serviceWorker.unregister();
