+++ src/reducers/index.js
@@ -1,0 +1,8 @@
+import { combineReducers } from 'redux';
+import bookReducer from './books';
+
+const rootReducer = combineReducers({
+  books: bookReducer,
+});
+
+export default rootReducer;
